Deleted DESCRIPTION Improved README.md, header guards and header includes Moved all files inside "/include" to "/include/untitled" Renamed a few of headers
