--- DESCRIPTION.txt
@@ -1,1 +1,0 @@
-Logger for C++ with namespace, class, struct etc.
