+++ README.md
@@ -7,7 +7,7 @@ A C++ library for my projects.
 Generic wrapper logger.
 
 ```cpp
-#include <untitled/logger.hh>
+#include <untitled/console/logger.hh>
 using Log = Untitled:Console::Logger;
 
 int main(void)
@@ -25,3 +25,14 @@ Features:
 ## UntitledMath
 
 Library for studying mathematical operations.
+
+```cpp
+#include <untitled/math/functions/square_root.hpp>
+using sqrt = Untitled::Math::square_root;
+int main()
+{
+    float x = 9;
+    float x_squared = sqrt(x);
+    return 0;
+}
+```
