+++ include/untitled/math/functions/convert.hpp
@@ -1,8 +1,8 @@
-#ifndef CONVERT_GUARD
-#define CONVERT_GUARD
+#ifndef UNTITLED_CONVERT_GUARD
+#define UNTITLED_CONVERT_GUARD
 namespace Untitled::Math
 {
     float degrees_to_radians(float number_in_degrees);
     float radians_to_degrees(float number_in_radians);
 } // namespace Untitled::Math
-#endif // CONVERT_GUARD
+#endif // UNTITLED_CONVERT_GUARD
