+++ include/untitled/math/functions/trigonometric_functions.hpp
@@ -1,9 +1,9 @@
-#ifndef TRIGONOMETRY_GUARD
-#define TRIGONOMETRY_GUARD
+#ifndef UNTITLED_TRIGONOMETRY_GUARD
+#define UNTITLED_TRIGONOMETRY_GUARD
 namespace Untitled::Math
 {
     float sine(float angle);
     float cosine(float angle);
     float tangent(float angle);
 } // namespace Untitled::Math
-#endif // TRIGONOMETRY_GUARD
+#endif // UNTITLED_TRIGONOMETRY_GUARD
