+++ include/untitled/utils/trigonometry.hpp
@@ -1,0 +1,5 @@
+#ifndef UNTITLED_TRIGONOMETRY_GUARD
+#define UNTITLED_TRIGONOMETRY_GUARD
+#include <untitled/math/functions/trigonometric_functions.hpp>
+#include <untitled/math/functions/square_root.hpp>
+#endif // UNTITLED_TRIGONOMETRY_GUARD
