+++ include/untitled/utils/vector.hpp
@@ -1,0 +1,5 @@
+#ifndef HEADER_VECTOR_GUARD
+#define HEADER_VECTOR_GUARD
+#include <untitled/math/objects/vector2d.hpp>
+#include <untitled/math/objects/vector3d.hpp>
+#endif // HEADER_VECTOR_GUARD
