--- include/utils/vector.hpp
@@ -1,5 +1,0 @@
-#ifndef VECTOR_GUARD
-#define VECTOR_GUARD
-#include "math/objects/vector2d.hpp"
-#include "math/objects/vector3d.hpp"
-#endif // VECTOR_GUARD
